net: add init controller functionality to prepare networking.
When the driver starts, the host machine network requires minor
configuration in order to accept VM workloads. This work currently
configures new iptables chains and corresponding jump rules.

The new functionality is built into the net interface which can be
called from the driver when required. The controller implements
this requirement.

The new tests need to be run as root, therefore the CI workflow
has been updated to allow this.