net: add teardown spec to task state for task recovery.
Adding the teardown spec to the task state for persistence via
the client allows recovering a tasks network subsystem correctly
in the event of a latter destroy. Without this, the network
controller would need to regenerate the assumed iptables rules,
which could be wrong due to changes between failure and recovery.