net: correctly return and handle started response with no network.
It is possible operators do not configure a network for a VM
driver task. In this case, the plugin should correctly handle the
response from the network controller.

Along with the main fix, the change updates the controller
responses, so that an object is returned when there isn't an
error. This is because returning nil,nil is a weird pattern in Go
and catches people out; even the person that wrote it.