fingerprint: fix unit and adjust attribute names for clarity
The `driver.virt.inactive` attribute was being incorrectly suffixed with
"bytes". The names of the fingerprinted fields for the number of active and
inactive domains are potentially confusing, as they don't refer to what about
the driver is active or inactive. Adjust the name to include `domains` in such a
way that active and inactive domains will sort together.