fingerprint: remove active/inactive domains attributes (#65)
The `driver.virt.active` and `driver.virt.inactive` attributes reflect the
number of VMs running. But this data is already available to the Nomad servers
via `Node.UpdateStatus` RPCs. Sending it as part of a periodic fingerprint will
cause Nomad client nodes to update their fingerprints regularly, kicking off new
evaluations for all the jobs on the node.