net: release DHCP lease on teardown
When performing the network teardown, make a best effort to remove
the DHCP lease that was assigned to the VM. There is no direct API
for removing a DHCP lease, but it can be accomplished by constructing
a custom release packet and sending it to the DHCP server. Since
the lease removal is only best effort, any errors that are encountered
are simply logged.