testing: add mocks and allow skipping iptables
This adds non-static mocks for the libvirt.ConnectShim and
libvirt.ConnectNetworkShim interfaces as well as a non-static
mock for IPTables. Mocks have been relocated into a common
and isolated location. Existing static mocks for both
ConnectShim and ConnectNetworkShim interfaces have been retained
with tests updated to import them for their new location.

An additional helper is added to allow tests to be skipped if
iptables is not available either due to the process not being
run as root or iptables not being available.