net: perform DHCP lookup using interface MAC address (#131)
This adds some plumbing for getting network interface information
for domains. The network interface MAC address can be extracted
and then used for IP address discovery via DHCP lease information.

Include the domain name and the hostname in the VMStartedBuildRequest
and use the Hostname field when referencing the hostname. Within
the DHCP IP lookup helper, check that the lease hostname matches
if it is set.