Testing updates (#151)
* testing: adjust test paths and add root checks where required

Update tests to use `testing.TempDir` where temporary directories
are required. Within cloudinit tests, move paths within the temp
directory so writes aren't directed at the root of the filesystem.

Add a `testutil` package which currently just has a root user
check. Add root user check to tests which require root.

* testing: add mocks and allow skipping iptables

This adds non-static mocks for the libvirt.ConnectShim and
libvirt.ConnectNetworkShim interfaces as well as a non-static
mock for IPTables. Mocks have been relocated into a common
and isolated location. Existing static mocks for both
ConnectShim and ConnectNetworkShim interfaces have been retained
with tests updated to import them for their new location.

An additional helper is added to allow tests to be skipped if
iptables is not available either due to the process not being
run as root or iptables not being available.

* testing: add mocked tests where elevated privileges required

For all tests that require elevated privileges, add complementary
mocked tests. This allows for testing expected behavior while not
running with elevated privileges (during development) but still
includes the full testing when privileges are elevated.